PhaseColors

Documentation for PhaseColors.

PhaseColors.canonicalphase

PhaseColors.phasecolor

PhaseColors.portrait

PhaseColors.portrait

PhaseColors.canonicalphase — Method

canonicalphase(z)

returns the phase of complex number, normalised from -pi,pi to 0,1

PhaseColors.phasecolor — Method

phasecolor(cm::ColorScheme)

returns a mapping from a number x to an RGB color corresponding to the colorscheme cm and the phase of x.

PhaseColors.portrait — Method

portrait(z,cm::ColorScheme)

colors an array z according angle(z) using the colormap cm

PhaseColors.portrait — Method

portrait(z,cm::Function)

colors an array z according angle(z) using the color function cm